requirements.txt
aiohttp==3.6.2
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.4.0
async-timeout==3.0.1
atomicwrites==1.1.5
attrs==19.3.0
basemap==1.1.0
beautifulsoup4==4.9.1
Brlapi==0.6.6
bs4==0.0.1
cairocffi==0.8.0
certifi==2020.4.5.1
cffi==1.11.5
chardet==3.0.4
colorspacious==1.1.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
discord==1.0.1
discord.py==1.3.3
distro-info===0.18ubuntu0.18.04.1
greenlet==0.4.12
httplib2==0.9.2
idna==2.9
idna-ssl==1.1.0
isort==4.3.21
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
mock==2.0.0
more-itertools==4.2.0
msgpack==0.5.6
multidict==4.7.6
neovim==0.2.6
netifaces==0.10.4
nose==1.3.7
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
pbr==3.1.1
pexpect==4.2.1
Pillow==5.1.0
pluggy==0.6.0
ply==3.11
protobuf==3.0.0
py==1.5.2
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.5.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyproj==1.9.5.1
pyRFC3339==1.0
pyshp==1.2.12
pytest==3.6.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.13.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
soupsieve==2.0.1
sphinx-gallery==0.1.13
system-service==0.3
systemd-python==234
toml==0.10.0
tornado==4.5.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
wadllib==1.3.2
websockets==8.1
wrapt==1.12.1
xkit==0.0.0
yarl==1.4.2
zope.interface==4.3.2
